refactor: split WrRecColl from coll-info to coll.js
